// common/spacetime/types/constraints.ts
type BaseTimeUnit =
  | 'millisecond'
  | 'second'
  | 'minute'
  | 'hour'
  | 'day'
  | 'week'
  | 'month'
  | 'year'
  | 'date';

export type TimeUnit = BaseTimeUnit | `${BaseTimeUnit}s`;

export type Format =
  | 'day'
  | 'day-short'
  | 'day-number'
  | 'day-pad'
  | 'date'
  | 'date-pad'
  | 'month'
  | 'month-short'
  | 'month-number'
  | 'month-pad'
  | 'year'
  | 'year-short'
  | 'time'
  | 'time-24'
  | 'hour'
  | 'hour-pad'
  | 'hour-24'
  | 'hour-24-pad'
  | 'minute'
  | 'minute-pad'
  | 'second'
  | 'second-pad'
  | 'millisecond'
  | 'ampm'
  | 'offset'
  | 'numeric'
  | 'numeric-us'
  | 'numeric-uk'
  | 'mm/dd'
  | 'iso'
  | 'iso-short'
  | 'iso-utc'
  | 'nice'
  | 'nice-year'
  | 'nice-day'
  | 'nice-full';